Trampoline and checkpoint update
Trampolines now animated.
Trampolines have a dotted line to show where they will be placed before creation.
Trampolines now disappear after use.

Checkpoint system added that includes prefabs that can be drag and dropped into the level to act as a new checkpoint or as a point that will reset the player if they enter it.